+++ bruteforce/tm_avx_16_test/tm_avx_16_test.cpp
@@ -81,7 +81,7 @@ void tm_avx_16_test::process_test_case(uint8 * test_case, uint16 * rng_seed, int
 
 	for (int i = 0; i < 128; i++)
 	{
-		test_case[i] = working_code[i];
+		test_case[i] = (uint8)working_code[i];
 	}
 
 	*rng_seed = this->rng_seed;
