+++ bruteforce/tm_avx_in_cpu_test/tm_avx_in_cpu_test.cpp
@@ -47,7 +47,7 @@ void tm_avx_intrinsics2_test::process_test_case(uint8 * test_case, uint16 * rng_
 
 	for (int i = 0; i < 128; i++)
 	{
-		test_case[i] = working_code[i];
+		test_case[i] = (uint8)working_code[i];
 	}
 
 	*rng_seed = this->rng_seed;
